make "count" return an "NSUInteger" rather than an "unsigned int" as in the original method signature (thanks to Sebastian Reitenbach)

+++ sope-gdl1/GDLAccess/EOPrimaryKeyDictionary.m
@@ -104,7 +104,7 @@
     return [key isEqual:_key] ? value : nil;
 }
 
-- (unsigned int)count {
+- (NSUInteger)count {
     return 1;
 }
 - (BOOL)isNotEmpty {
